backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
paho-mqtt==1.6.1
influxdb-client==1.38.0
PyYAML==6.0.1
python-multipart==0.0.6
aiohttp==3.9.1
